Add keyframe for changing default words
