+++ website/src/lib/components/index.ts
@@ -14,4 +14,4 @@ export { verse1Keyframes } from "./verse1/keyframe.svelte";
 export { verse2Keyframes } from "./verse2/keyframe";
 export { verse3Keyframes } from "./verse3/keyframe";
 export { verse4Keyframes } from "./verse4/keyframe";
-export { verse5Keyframes } from "./test/keyframe";
+export { verse5Keyframes } from "./test/keyframe.svelte";
