--- website/src/lib/components/test/keyframe.ts
@@ -1,28 +1,0 @@
-import { type Keyframe } from "$lib/keyframe.svelte";
-
-const verse = 5;
-/**
- * The keyframes for verse test.
- */
-export const verse5Keyframes: Keyframe[] = [
-    {
-        verse: verse,
-        bolded: [0],
-        toRun: []
-    },
-    {
-        verse: verse,
-        bolded: [1],
-        toRun: []
-    },
-    {
-        verse: verse,
-        bolded: [2],
-        toRun: []
-    },
-    {
-        verse: verse,
-        bolded: [3],
-        toRun: []
-    }
-];
