+++ website/src/lib/components/test/word_chart.svelte
@@ -4,6 +4,7 @@ map component used in verse 3 for visualizing geographic data
 <script lang="ts">
     import { homeSimilarityData } from "$lib/data";
     import { extent, scaleLinear } from "d3";
+    import { defaultWords } from "./keyframe.svelte";
     import cloud from "d3-cloud";
 
     // chart limits
@@ -18,9 +19,7 @@ map component used in verse 3 for visualizing geographic data
     const wordScale = scaleLinear().domain(similarityExtent).range([10, 100]);
 
     // interactivity
-    let defaultWords = $state([""]);
     let hoveredWord = $state("");
-    $inspect(hoveredWord).with(console.log);
 
     // data
     const data = $derived(
@@ -84,14 +83,14 @@ map component used in verse 3 for visualizing geographic data
 >
     <!-- title -->
     <div
-        class="absolute top-0 left-1/2 z-10 -translate-x-1/2 rounded-md bg-white/70 text-center text-xl"
+        class="absolute left-1/2 top-0 z-10 -translate-x-1/2 rounded-md bg-white/70 text-center text-xl"
     >
         What Does Home Mean to You?
     </div>
 
     <!-- info tooltip -->
     <div
-        class="tooltip tooltip-left absolute right-5 bottom-5 z-10 h-fit w-fit rounded-full p-0 hover:cursor-pointer"
+        class="tooltip tooltip-left absolute bottom-5 right-5 z-10 h-fit w-fit rounded-full p-0 hover:cursor-pointer"
         data-tip={"Hover over a word to see cosine similarity with 'Home'." +
             "\nWords collected from Habitat for Humanity's 'What does home mean to you' page, and cosine similarity calculated using API Ninjas' Text Similarity API."}
     >
